Move roxie row allocator into roxiehelper
